requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
matplotlib>=3.3.2
seaborn>=0.11.0
scikit-learn>=0.24.2
tqdm>=4.50.2
PyYAML>=5.4.1
einops>=0.3.0
thop>=0.0.31
wandb>=0.12.0
pillow>=8.2.0
pandas>=1.2.4
tensorboard>=2.5.0 